+++ inference-chain/app/upgrades/v0_2_13/upgrades.go
@@ -84,7 +84,6 @@ var devshardAllowedCreatorAddressesToAdd = []string{
 const (
 	MaxEscrowsPerEpoch uint32 = 500_000
 	MaxNonce           uint32 = 1_000_000
-
 	// Block window after the upgrade in which confirmation PoC is skipped.
 	// Same value as v0.2.10; covers the rest of the upgrade epoch on mainnet.
 	GraceUpgradeProtectionWindow int64 = 10000
@@ -139,16 +138,10 @@ func CreateUpgradeHandler(
 		if err := setDevshardEscrowParams(ctx, k); err != nil {
 			return nil, err
 		}
-		if err := backfillDevshardEscrowFees(ctx, k); err != nil {
-			return nil, err
-		}
 		if err := setDevshardAllowedCreatorAddresses(ctx, k); err != nil {
 			return nil, err
 		}
 		if err := setDevshardApprovedVersions(ctx, k); err != nil {
-			return nil, err
-		}
-		if err := setDevshardApprovedV2Version(ctx, k); err != nil {
 			return nil, err
 		}
 		if err := backfillConfirmationWeightScales(ctx, k); err != nil {
@@ -247,143 +240,13 @@ func setDevshardEscrowParams(ctx context.Context, k keeper.Keeper) error {
 	params.DevshardEscrowParams.MaxEscrowsPerEpoch = MaxEscrowsPerEpoch
 	params.DevshardEscrowParams.MaxNonce = MaxNonce
 	params.DevshardEscrowParams.DevshardRequestsEnabled = types.DefaultDevshardRequestsEnabled
-
-	// Initialize DefaultSealGraceNonces for existing chains. The field is new
-	// in v0.2.13: a fresh-genesis chain seeds it from
-	// types.DefaultDevshardSealGraceNonces(GroupSize), but an existing chain
-	// upgrading from <=v0.2.12 deserializes it to 0. Leaving 0 would diverge
-	// from fresh-genesis state and break any consumer that reads the chain
-	// default (e.g. governance UIs, AppExport parsers, future on-chain
-	// callers). We backfill the canonical value here, using the existing
-	// GroupSize when present so the derivation matches genesis exactly.
-	if params.DevshardEscrowParams.DefaultSealGraceNonces == 0 {
-		groupSize := params.DevshardEscrowParams.GroupSize
-		if groupSize == 0 {
-			groupSize = types.DefaultDevshardGroupSize
-		}
-		params.DevshardEscrowParams.DefaultSealGraceNonces = types.DefaultDevshardSealGraceNonces(groupSize)
-	}
-	if params.DevshardEscrowParams.DefaultInferenceClearGraceSeconds == 0 {
-		params.DevshardEscrowParams.DefaultInferenceClearGraceSeconds = types.DefaultDevshardInferenceClearGraceSeconds
-	}
-
-	// v0.2.13 adds fee + bind-time governance fields. Existing chains decode
-	// these as zero (proto3 default) and would fail DevshardEscrowParams.Validate
-	// because refusal_timeout/execution_timeout/vote_threshold_factor must be > 0.
-	// Backfill zero-valued fields with canonical defaults; non-zero values left
-	// in place so any pre-existing governance override survives the upgrade.
-	if params.DevshardEscrowParams.CreateDevshardFee == 0 {
-		params.DevshardEscrowParams.CreateDevshardFee = types.DefaultDevshardCreateDevshardFee
-	}
-	if params.DevshardEscrowParams.FeePerNonce == 0 {
-		params.DevshardEscrowParams.FeePerNonce = types.DefaultDevshardFeePerNonce
-	}
-	if params.DevshardEscrowParams.RefusalTimeout == 0 {
-		params.DevshardEscrowParams.RefusalTimeout = types.DefaultDevshardRefusalTimeout
-	}
-	if params.DevshardEscrowParams.ExecutionTimeout == 0 {
-		params.DevshardEscrowParams.ExecutionTimeout = types.DefaultDevshardExecutionTimeout
-	}
-	if params.DevshardEscrowParams.ValidationRate == 0 {
-		params.DevshardEscrowParams.ValidationRate = types.DefaultDevshardValidationRate
-	}
-	if params.DevshardEscrowParams.VoteThresholdFactor == 0 {
-		params.DevshardEscrowParams.VoteThresholdFactor = types.DefaultDevshardVoteThresholdFactor
-	}
-
 	if err := k.SetParams(ctx, params); err != nil {
 		return err
 	}
 	k.LogInfo("set devshard escrow params", types.Upgrades,
 		"max_escrows_per_epoch", MaxEscrowsPerEpoch,
 		"max_nonce", MaxNonce,
-		"devshard_requests_enabled", params.DevshardEscrowParams.DevshardRequestsEnabled,
-		"default_seal_grace_nonces", params.DevshardEscrowParams.DefaultSealGraceNonces,
-		"default_inference_clear_grace_seconds", params.DevshardEscrowParams.DefaultInferenceClearGraceSeconds,
-		"create_devshard_fee", params.DevshardEscrowParams.CreateDevshardFee,
-		"fee_per_nonce", params.DevshardEscrowParams.FeePerNonce,
-		"refusal_timeout", params.DevshardEscrowParams.RefusalTimeout,
-		"execution_timeout", params.DevshardEscrowParams.ExecutionTimeout,
-		"validation_rate", params.DevshardEscrowParams.ValidationRate,
-		"vote_threshold_factor", params.DevshardEscrowParams.VoteThresholdFactor,
-	)
-	return nil
-}
-
-// backfillDevshardEscrowFees populates the new per-escrow fee snapshot fields
-// (create_devshard_fee, fee_per_nonce) on DevshardEscrow rows created before
-// v0.2.13. Pre-upgrade rows decode both fields as zero; settlement and any
-// downstream consumer expect a non-zero per-escrow snapshot matching the
-// governance values active when the escrow was created. Since pre-upgrade
-// escrows have no historical record of those values, we backfill with the
-// current (just-migrated) DevshardEscrowParams values, which match what a
-// fresh CreateDevshardEscrow would write today. Rows that already carry a
-// non-zero snapshot (post-upgrade creates, or future re-runs of this handler)
-// are left untouched.
-func backfillDevshardEscrowFees(ctx context.Context, k keeper.Keeper) error {
-	params, err := k.GetParams(ctx)
-	if err != nil {
-		return err
-	}
-	if params.DevshardEscrowParams == nil {
-		k.LogInfo("backfill devshard escrow fees skipped: devshard escrow params missing", types.Upgrades)
-		return nil
-	}
-	createFee := params.DevshardEscrowParams.CreateDevshardFee
-	feePerNonce := params.DevshardEscrowParams.FeePerNonce
-
-	// Keep memory bounded during Walk: record only escrow IDs that need a backfill,
-	// then re-read and update each row. This avoids buffering full escrow objects
-	// while still preserving two-phase iteration (no writes during Walk).
-	//
-	// Important: we intentionally do not call SetDevshardEscrow inside Walk.
-	// collections.Map iteration is safest when mutation happens after traversal,
-	// so future SDK/internal iterator changes cannot invalidate traversal state.
-	var updateIDs []uint64
-	if err := k.DevshardEscrows.Walk(ctx, nil, func(_ uint64, escrow types.DevshardEscrow) (bool, error) {
-		if escrow.CreateDevshardFee != 0 && escrow.FeePerNonce != 0 {
-			return false, nil
-		}
-		updateIDs = append(updateIDs, escrow.Id)
-		return false, nil
-	}); err != nil {
-		return fmt.Errorf("walk devshard escrows for fee backfill: %w", err)
-	}
-
-	for _, id := range updateIDs {
-		escrow, found := k.GetDevshardEscrow(ctx, id)
-		if !found {
-			return fmt.Errorf("get devshard escrow %d during fee backfill: not found", id)
-		}
-		if escrow.CreateDevshardFee == 0 {
-			escrow.CreateDevshardFee = createFee
-		}
-		if escrow.FeePerNonce == 0 {
-			escrow.FeePerNonce = feePerNonce
-		}
-		if err := k.SetDevshardEscrow(ctx, escrow); err != nil {
-			return fmt.Errorf("set devshard escrow %d during fee backfill: %w", escrow.Id, err)
-		}
-	}
-	k.LogInfo("backfilled devshard escrow fees", types.Upgrades,
-		"updated", len(updateIDs),
-		"create_devshard_fee", createFee,
-		"fee_per_nonce", feePerNonce,
-	)
-	return nil
-}
-
-func setDevshardApprovedV2Version(ctx context.Context, k keeper.Keeper) error {
-	params, err := k.GetParams(ctx)
-	if err != nil {
-		return err
-	}
-	if params.DevshardEscrowParams == nil {
-		params.DevshardEscrowParams = types.DefaultDevshardEscrowParams()
-	}
-	if err := k.SetParams(ctx, params); err != nil {
-		return err
-	}
+		"devshard_requests_enabled", types.DefaultDevshardRequestsEnabled)
 	return nil
 }
