+++ inference-chain/app/upgrades/v0_2_13/upgrades_test.go
@@ -98,186 +98,6 @@ func TestSetDevshardEscrowParamsEnablesRequests(t *testing.T) {
 	require.True(t, got.DevshardEscrowParams.DevshardRequestsEnabled)
 }
 
-func TestSetDevshardEscrowParams_BackfillsDefaultSealGraceNonces(t *testing.T) {
-	k, ctx, _ := keepertest.InferenceKeeperReturningMocks(t)
-
-	params, err := k.GetParams(ctx)
-	require.NoError(t, err)
-	require.NotNil(t, params.DevshardEscrowParams)
-	// Simulate pre-v0.2.13 state where the new field is zero on disk.
-	params.DevshardEscrowParams.DefaultSealGraceNonces = 0
-	require.NoError(t, k.SetParams(ctx, params))
-
-	require.NoError(t, setDevshardEscrowParams(ctx, k))
-
-	got, err := k.GetParams(ctx)
-	require.NoError(t, err)
-	require.NotNil(t, got.DevshardEscrowParams)
-	expected := inferencetypes.DefaultDevshardSealGraceNonces(got.DevshardEscrowParams.GroupSize)
-	require.Equal(t, expected, got.DevshardEscrowParams.DefaultSealGraceNonces)
-	require.Equal(t, MaxEscrowsPerEpoch, got.DevshardEscrowParams.MaxEscrowsPerEpoch)
-	require.Equal(t, MaxNonce, got.DevshardEscrowParams.MaxNonce)
-	require.Equal(t, inferencetypes.DefaultDevshardInferenceClearGraceSeconds, got.DevshardEscrowParams.DefaultInferenceClearGraceSeconds)
-}
-
-func TestSetDevshardEscrowParams_BackfillsDefaultInferenceClearGraceSeconds(t *testing.T) {
-	k, ctx, _ := keepertest.InferenceKeeperReturningMocks(t)
-
-	params, err := k.GetParams(ctx)
-	require.NoError(t, err)
-	require.NotNil(t, params.DevshardEscrowParams)
-	params.DevshardEscrowParams.DefaultInferenceClearGraceSeconds = 0
-	require.NoError(t, k.SetParams(ctx, params))
-
-	require.NoError(t, setDevshardEscrowParams(ctx, k))
-
-	got, err := k.GetParams(ctx)
-	require.NoError(t, err)
-	require.Equal(t, inferencetypes.DefaultDevshardInferenceClearGraceSeconds, got.DevshardEscrowParams.DefaultInferenceClearGraceSeconds)
-}
-
-// TestSetDevshardEscrowParams_BackfillsPhase4Fields covers the six fields added
-// in v0.2.13 (create_devshard_fee, fee_per_nonce, refusal_timeout,
-// execution_timeout, validation_rate, vote_threshold_factor). On an existing
-// chain these decode as zero and would fail Validate; the migration must seed
-// canonical defaults so the post-upgrade params validate.
-func TestSetDevshardEscrowParams_BackfillsPhase4Fields(t *testing.T) {
-	k, ctx, _ := keepertest.InferenceKeeperReturningMocks(t)
-
-	params, err := k.GetParams(ctx)
-	require.NoError(t, err)
-	require.NotNil(t, params.DevshardEscrowParams)
-	params.DevshardEscrowParams.CreateDevshardFee = 0
-	params.DevshardEscrowParams.FeePerNonce = 0
-	params.DevshardEscrowParams.RefusalTimeout = 0
-	params.DevshardEscrowParams.ExecutionTimeout = 0
-	params.DevshardEscrowParams.ValidationRate = 0
-	params.DevshardEscrowParams.VoteThresholdFactor = 0
-	// SetParams marshals without invoking Validate(); this simulates a v0.2.12
-	// on-disk state where the new fields decode as zero.
-	require.NoError(t, k.SetParams(ctx, params))
-
-	require.NoError(t, setDevshardEscrowParams(ctx, k))
-
-	got, err := k.GetParams(ctx)
-	require.NoError(t, err)
-	require.Equal(t, inferencetypes.DefaultDevshardCreateDevshardFee, got.DevshardEscrowParams.CreateDevshardFee)
-	require.Equal(t, inferencetypes.DefaultDevshardFeePerNonce, got.DevshardEscrowParams.FeePerNonce)
-	require.Equal(t, inferencetypes.DefaultDevshardRefusalTimeout, got.DevshardEscrowParams.RefusalTimeout)
-	require.Equal(t, inferencetypes.DefaultDevshardExecutionTimeout, got.DevshardEscrowParams.ExecutionTimeout)
-	require.Equal(t, inferencetypes.DefaultDevshardValidationRate, got.DevshardEscrowParams.ValidationRate)
-	require.Equal(t, inferencetypes.DefaultDevshardVoteThresholdFactor, got.DevshardEscrowParams.VoteThresholdFactor)
-	require.NoError(t, got.DevshardEscrowParams.Validate())
-}
-
-func TestSetDevshardEscrowParams_PreservesExistingPhase4Fields(t *testing.T) {
-	k, ctx, _ := keepertest.InferenceKeeperReturningMocks(t)
-
-	params, err := k.GetParams(ctx)
-	require.NoError(t, err)
-	require.NotNil(t, params.DevshardEscrowParams)
-	params.DevshardEscrowParams.CreateDevshardFee = 42_000
-	params.DevshardEscrowParams.FeePerNonce = 7
-	params.DevshardEscrowParams.RefusalTimeout = 123
-	params.DevshardEscrowParams.ExecutionTimeout = 4567
-	params.DevshardEscrowParams.ValidationRate = 8800
-	params.DevshardEscrowParams.VoteThresholdFactor = 77
-	require.NoError(t, k.SetParams(ctx, params))
-
-	require.NoError(t, setDevshardEscrowParams(ctx, k))
-
-	got, err := k.GetParams(ctx)
-	require.NoError(t, err)
-	require.Equal(t, uint64(42_000), got.DevshardEscrowParams.CreateDevshardFee)
-	require.Equal(t, uint64(7), got.DevshardEscrowParams.FeePerNonce)
-	require.Equal(t, int64(123), got.DevshardEscrowParams.RefusalTimeout)
-	require.Equal(t, int64(4567), got.DevshardEscrowParams.ExecutionTimeout)
-	require.Equal(t, uint32(8800), got.DevshardEscrowParams.ValidationRate)
-	require.Equal(t, uint32(77), got.DevshardEscrowParams.VoteThresholdFactor)
-}
-
-// TestBackfillDevshardEscrowFees verifies that pre-v0.2.13 escrow rows (with
-// zero create_devshard_fee and fee_per_nonce) are populated with the current
-// DevshardEscrowParams values, while rows that already have a non-zero snapshot
-// are left untouched.
-func TestBackfillDevshardEscrowFees(t *testing.T) {
-	k, ctx, _ := keepertest.InferenceKeeperReturningMocks(t)
-
-	params, err := k.GetParams(ctx)
-	require.NoError(t, err)
-	require.NotNil(t, params.DevshardEscrowParams)
-	params.DevshardEscrowParams.CreateDevshardFee = 9_000
-	params.DevshardEscrowParams.FeePerNonce = 250
-	require.NoError(t, k.SetParams(ctx, params))
-
-	legacy := &inferencetypes.DevshardEscrow{
-		Creator: "gonka1legacy",
-		Amount:  100,
-		Slots:   []string{"s"},
-	}
-	_, err = k.StoreDevshardEscrow(ctx, legacy, 1)
-	require.NoError(t, err)
-
-	partial := &inferencetypes.DevshardEscrow{
-		Creator:           "gonka1partial",
-		Amount:            200,
-		Slots:             []string{"s"},
-		CreateDevshardFee: 1_111,
-	}
-	_, err = k.StoreDevshardEscrow(ctx, partial, 2)
-	require.NoError(t, err)
-
-	fresh := &inferencetypes.DevshardEscrow{
-		Creator:           "gonka1fresh",
-		Amount:            300,
-		Slots:             []string{"s"},
-		CreateDevshardFee: 5_000,
-		FeePerNonce:       500,
-	}
-	_, err = k.StoreDevshardEscrow(ctx, fresh, 3)
-	require.NoError(t, err)
-
-	require.NoError(t, backfillDevshardEscrowFees(ctx, k))
-
-	gotLegacy, found := k.GetDevshardEscrow(ctx, 1)
-	require.True(t, found)
-	require.Equal(t, uint64(9_000), gotLegacy.CreateDevshardFee)
-	require.Equal(t, uint64(250), gotLegacy.FeePerNonce)
-
-	gotPartial, found := k.GetDevshardEscrow(ctx, 2)
-	require.True(t, found)
-	require.Equal(t, uint64(1_111), gotPartial.CreateDevshardFee, "non-zero create_devshard_fee must be preserved")
-	require.Equal(t, uint64(250), gotPartial.FeePerNonce, "zero fee_per_nonce must be backfilled")
-
-	gotFresh, found := k.GetDevshardEscrow(ctx, 3)
-	require.True(t, found)
-	require.Equal(t, uint64(5_000), gotFresh.CreateDevshardFee)
-	require.Equal(t, uint64(500), gotFresh.FeePerNonce)
-}
-
-func TestBackfillDevshardEscrowFees_NoEscrowsIsNoOp(t *testing.T) {
-	k, ctx, _ := keepertest.InferenceKeeperReturningMocks(t)
-
-	require.NoError(t, backfillDevshardEscrowFees(ctx, k))
-}
-
-func TestSetDevshardEscrowParams_PreservesExistingDefaultSealGraceNonces(t *testing.T) {
-	k, ctx, _ := keepertest.InferenceKeeperReturningMocks(t)
-
-	params, err := k.GetParams(ctx)
-	require.NoError(t, err)
-	require.NotNil(t, params.DevshardEscrowParams)
-	const customGrace uint32 = 12345
-	params.DevshardEscrowParams.DefaultSealGraceNonces = customGrace
-	require.NoError(t, k.SetParams(ctx, params))
-
-	require.NoError(t, setDevshardEscrowParams(ctx, k))
-
-	got, err := k.GetParams(ctx)
-	require.NoError(t, err)
-	require.Equal(t, customGrace, got.DevshardEscrowParams.DefaultSealGraceNonces)
-}
-
 func TestAddDevshardAllowedCreatorAddressesAppendsOnlyNewAddresses(t *testing.T) {
 	k, ctx, _ := keepertest.InferenceKeeperReturningMocks(t)
