+++ inference-chain/app/upgrades/v0_2_14/constants.go
@@ -1,0 +1,3 @@
+package v0_2_14
+
+const UpgradeName = "v0.2.14"
